manufacturer and reseler api modification

+++ routes/api.php
@@ -19,9 +19,10 @@ Route::prefix('manufacturers')->group(function () {
     // Route::post('/basic-info', [ManufacturerController::class, 'storeBasicInfo']);
     // Route::post('/company-info', [ManufacturerController::class, 'storeCompanyInfo']);
     // Route::post('/confirm', [ManufacturerController::class, 'confirm']);
+    Route::get('', [ManufacturerController::class, 'getAllManufacturers']);
     Route::post('/store', [ManufacturerController::class, 'storeManufacturer']);
     Route::get('/{user_id}', [ManufacturerController::class, 'getManufacturerDetails']);
-    Route::delete('/delete', [ManufacturerController::class, 'deleteManufacturer']);
+    Route::delete('/delete/{user_id}', [ManufacturerController::class, 'deleteManufacturer']);
 });
 
 //resellers
@@ -29,9 +30,10 @@ Route::prefix('resellers')->group(function () {
     // Route::post('/basic-info', [ResellerController::class, 'storeBasicInfo']);
     // Route::post('/company-info', [ResellerController::class, 'storeCompanyInfo']);
     // Route::post('/confirm', [ResellerController::class, 'confirm']);
+    Route::get('', [ResellerController::class, 'getAllResellers']);
     Route::post('/store', [ResellerController::class, 'storeReseller']);
     Route::get('/{user_id}', [ResellerController::class, 'getResellerDetails']);
-    Route::delete('/delete', [ResellerController::class, 'deleteReseller']);
+    Route::delete('/delete/{user_id}', [ResellerController::class, 'deleteReseller']);
 });
 
 //blogs
